% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olr_function.R
\name{olr}
\alias{olr}
\alias{olrmodels}
\alias{olrformulas}
\alias{olrformulaorder}
\alias{adjr2list}
\alias{r2list}
\title{olr: Optimal Linear Regression}
\usage{
olr(datasetname, resvarname, expvarnames, adjr2 = TRUE)

olrmodels(datasetname, resvarname, expvarnames)

olrformulas(datasetname, resvarname, expvarnames)

olrformulaorder(datasetname, resvarname, expvarnames)

adjr2list(datasetname, resvarname, expvarnames)

r2list(datasetname, resvarname, expvarnames)
}
\arguments{
\item{datasetname}{is defined by the user and points to the name of the dataset that is being used.}

\item{resvarname}{the response variable name defined as a string. For example. It represents a header in the data table.}

\item{expvarnames}{is a variable that you must define as a list before your function. Place desired headers from the data table in here as a list.}

\item{adjr2}{adjr2 = TRUE means you want to return the regression summary for the maximum adjusted R-squared term. adjr2 = FALSE means you want to return the regression summary for the maximum R-squared term.}
}
\value{
The regression summary for the adjusted R-squared or the R-squared, specified with TRUE or FALSE in the olr function.
}
\description{
The main function olr runs all of the possible linear regression equation combinations, which are all of the
combinations of dependent variables respect to the independent variable. In essence, the olr function
returns the best fit linear regression model. The user can prompt the olr function to either return the
best fit statistical summary of either the greatest adjusted R-squared or R-squared term.
}
\details{
Complimentary functions below follow this format: function(datasetname, resvarname, expvarnames) \cr \cr
olrmodels: returns the list of models accompanied by the coefficients. After typing in olrmodels(datasetname, resvarname, expvarnames) put the number you want to see in the brackets on the right of the comma [,]. For example, olrmodels(datasetname, resvarname, expvarnames)[,1] \cr \cr
olrformulas: returns the list of olr formulas \cr \cr
olrformulasorder: returns the formulas with the predictors (dependent variables) in ascending order \cr \cr
adjr2list: list of the adjusted R-squared terms \cr \cr
r2list: list of the R-squared terms \cr \cr

A Python version is available at <https://pypi.org/project/olr/>.
}
\examples{
\dontrun{
For this example, 'Import Dataset' from this location: ../olr/insta/extdata

Or use the function to find the location:
system.file("extdata", "oildata.csv", package = "olr", mustWork = TRUE)

Go to "Environment" and go to "Import Dataset". Then choose "From Text (base)...".
Afterwards, find the file in the directory. After clicking the file, choose "Yes" for "Headers"
Then, click on "Import"

oildata should now be listed as one of the global variables.

datasetname <- oildata
resvarname <- 'OilPrices'
expvarnames <- c('SP500', 'RigCount', 'API', 'Field_Production', 'RefinerNetInput',
'OperableCapacity', 'Imports', 'StocksExcludingSPR')
olr(datasetname, resvarname, expvarnames, adjr2 = TRUE)
}
}
\keyword{olr}
